Renaming API files. Moving supabase file
